+++ src/main/java/com/olpasa/service/impl/CriterioCalidadServiceImpl.java
@@ -38,7 +38,7 @@ public class CriterioCalidadServiceImpl implements ICriterioCalidadService{
 	}
 
 	@Override
-	public void eliminar(Long id) {
+	public void eliminar(Integer id) {
 		//sectorRepo.deleteById(id);
 	}
