+++ src/main/java/com/olpasa/service/impl/CuentaBancariaSericeImpl.java
@@ -38,10 +38,9 @@ public class CuentaBancariaSericeImpl implements ICuentaBancariaService{
 	}
 
 	@Override
-	public void eliminar(Long id) {
-		//sectorRepo.deleteById(id);
+	public void eliminar(Integer id) {
+
 	}
-	
-	
+
 	
 }
