+++ src/main/java/com/olpasa/service/impl/ProductoServiceImpl.java
@@ -37,7 +37,7 @@ public class ProductoServiceImpl implements IProductoService{
 	}
 
 	@Override
-	public void eliminar(Long id) {
+	public void eliminar(Integer id) {
 		//sectorRepo.deleteById(id);
 	}
